requirements.txt
python-telegram-bot[webhooks]>=21.0
psycopg2-binary
python-dotenv
flask
pytz
requests
hmac
hashlib